remove readonly move beaconstorage reduce package number
