+++ go/beacon_srv/internal/beaconstorage/BUILD.bazel
@@ -3,16 +3,16 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "beacon.go",
-        "revocation.go",
+        "policy.go",
+        "store.go",
     ],
-    importpath = "github.com/scionproto/scion/go/lib/beaconstorage/readonly",
+    importpath = "github.com/scionproto/scion/go/beacon_srv/internal/beaconstorage",
     visibility = ["//visibility:public"],
     deps = [
         "//go/lib/addr:go_default_library",
         "//go/lib/common:go_default_library",
         "//go/lib/ctrl/path_mgmt:go_default_library",
         "//go/lib/ctrl/seg:go_default_library",
-        "//go/lib/spath:go_default_library",
+        "//go/proto:go_default_library",
     ],
 )
