--- go/lib/beaconstorage/BUILD.bazel
@@ -1,15 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["common.go"],
-    importpath = "github.com/scionproto/scion/go/lib/beaconstorage",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//go/lib/addr:go_default_library",
-        "//go/lib/beaconstorage/beaconpol:go_default_library",
-        "//go/lib/beaconstorage/readonly:go_default_library",
-        "//go/lib/common:go_default_library",
-        "//go/proto:go_default_library",
-    ],
-)
