--- go/lib/beaconstorage/beaconpol/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["policy.go"],
-    importpath = "github.com/scionproto/scion/go/lib/beaconstorage/beaconpol",
-    visibility = ["//visibility:public"],
-    deps = ["//go/lib/addr:go_default_library"],
-)
